requirements.txt
google-cloud-vision
ipykernel
google-cloud-translate
Flask
flask-cors
Pillow